requirements.txt
langchain 
langchain-openai 
langchain-community 
langgraph 
langchain-chroma 
chromadb 
openai 
python-dotenv 
tiktoken